Refactor - move tree initializer to its own module. (#740)
