+++ src/lib.rs
@@ -7,6 +7,8 @@ pub mod config;
 mod contracts;
 mod database;
 mod ethereum;
+
+mod identity;
 pub mod identity_tree;
 pub mod prover;
 pub mod server;
